migrate models.py to folder
