+++ estate_agent/estate_agent/models/__init__.py
@@ -1,0 +1,2 @@
+from .property import Property
+from .property_image import PropertyImage
